Fix script to work with Ruby 2.6.8
When bootstrapping dotfiles on a new laptop, `asdf` hasn't been
installed yet, so we rely on the version of Ruby that's preinstalled on
macOS, which at the time of writing is 2.6.8. This doesn't support
`Symbol#end_with?` yet, so this commit fixes a bug so as not to rely on
that.